@@ -1,24 +1,25 @@
 {
   "name": "coriolis-readme-samples",
   "version": "1.0.0",
-  "dependencies": {
-    "@coriolis/coriolis": "*"
-  },
   "scripts": {
     "lint": "eslint .",
     "pretty": "prettier './**/{*.{js?(on),md},.*.{js?(on),md}}' --write && eslint . --fix",
-    "start": "node -r @babel/register cli.js"
+    "start": "node -r @babel/register cli.js",
+    "upgrade": "npm outdated --parseable --depth=0 | cut -d: -f4 | xargs npm install"
+  },
+  "dependencies": {
+    "@coriolis/coriolis": "file:../.."
   },
   "devDependencies": {
-    "@babel/core": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/register": "^7.10.4",
+    "@babel/core": "^7.16.5",
+    "@babel/preset-env": "^7.16.5",
+    "@babel/register": "^7.16.5",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.4.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-standard": "^4.0.1",
-    "prettier": "^2.0.5"
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-standard": "^5.0.0",
+    "prettier": "^2.5.1"
   }
 }
