@@ -2,34 +2,35 @@
   "name": "todos-svelte-example",
   "version": "1.0.0",
   "devDependencies": {
-    "node-sass": "^4.14.1",
     "npm-run-all": "^4.1.5",
-    "rollup": "^2.21.0",
+    "rollup": "^2.70.2",
     "rollup-plugin-commonjs": "^10.0.0",
-    "rollup-plugin-livereload": "^1.3.0",
+    "rollup-plugin-livereload": "^2.0.0",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-svelte": "^5.2.3",
-    "rollup-plugin-terser": "^6.1.0",
-    "svelte": "^3.24.0",
-    "svelte-preprocess": "^4.0.6"
+    "rollup-plugin-svelte": "^7.1.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "sass": "^1.50.1",
+    "svelte": "^3.47.0",
+    "svelte-preprocess": "^4.10.6"
   },
   "dependencies": {
-    "@coriolis/coriolis": "*",
-    "@coriolis/dev-tools": "*",
-    "@coriolis/parametered-projection": "*",
-    "@coriolis/coriolis-svelte": "*",
+    "@coriolis/coriolis": "file:../..",
+    "@coriolis/coriolis-svelte": "file:../../../coriolis-svelte",
+    "@coriolis/dev-tools": "file:../../../dev-tools",
+    "@coriolis/parametered-projection": "file:../../../parametered-projection",
     "@sveltejs/svelte-virtual-list": "^3.0.1",
-    "immer": "^7.0.5",
-    "sirv-cli": "^1.0.3"
+    "immer": "^9.0.12",
+    "sirv-cli": "^2.0.2"
   },
   "scripts": {
     "clear": "rm -rf node_modules public/bundle.css public/*.js public/*.map",
     "build": "rollup -c",
-    "autobuild": "rollup -c -w",
+    "autobuild": "rollup -c -w --watch.buildDelay 2000",
     "dev": "run-p start:dev autobuild",
     "dev:remote": "run-p start:dev:remote autobuild",
     "start": "sirv public --single",
-    "start:dev": "sirv public --single --dev",
-    "start:dev:remote": "npm run -s start:dev -- --host 0.0.0.0"
+    "start:dev": "sirv public --single --dev -p 8080",
+    "start:dev:remote": "npm run -s start:dev -- --host 0.0.0.0",
+    "upgrade": "npm outdated --parseable --depth=0 | cut -d: -f4 | xargs npm install"
   }
 }
