@@ -18,15 +18,18 @@ export default {
   },
   plugins: [
     svelte({
-      // enable run-time checks when not in production
-      dev: !production,
+      compilerOptions: {
+        // enable run-time checks when not in production
+        dev: !production,
+      },
       // we'll extract any component CSS out into
       // a separate file — better for performance
-      css: (css) => {
-        css.write('public/bundle.css')
-      },
+      // css: (css) => {
+      //   css.write('bundle.css')
+      // },
 
       preprocess: autoPreprocess(),
+      emitCss: false,
     }),
 
     // If you have external dependencies installed from
