@@ -1,1 +1,4 @@
-export const not = (func) => (...args) => !func(...args)
+export const not =
+  (func) =>
+  (...args) =>
+    !func(...args)
