@@ -1,5 +1,6 @@
 import { createIndex } from '../lib/map/objectIndex'
 
 export const fromReducer = createIndex(
-  (reducer) => ({ useState, useEvent }) => (useState(), useEvent(), reducer),
+  (reducer) =>
+    ({ useState, useEvent }) => (useState(), useEvent(), reducer),
 ).get
