@@ -1,53 +1,52 @@
-const createSubsetGetter = (index, getNotYetIndexed) => (
-  indexed,
-  key,
-  rest,
-) => {
-  let subset = indexed.subset
+const createSubsetGetter =
+  (index, getNotYetIndexed) => (indexed, key, rest) => {
+    let subset = indexed.subset
 
-  if (!subset) {
-    subset = createIndex((...args) => getNotYetIndexed(key, ...args))
+    if (!subset) {
+      subset = createIndex((...args) => getNotYetIndexed(key, ...args))
+
+      index.set(key, {
+        ...indexed,
+        subset,
+      })
+    }
+
+    return subset.get(...rest)
+  }
+
+const createValueGetter =
+  (
+    index,
+    getNotYetIndexed,
+    getSubset = createSubsetGetter(index, getNotYetIndexed),
+  ) =>
+  (key, ...rest) => {
+    const indexed = index.get(key) || {}
+
+    if (rest.length) {
+      return getSubset(indexed, key, rest)
+    }
+
+    if (indexed.loading) {
+      throw new Error('Cycle index referencing is not possible')
+    }
+
+    if ('value' in indexed) {
+      return indexed.value
+    }
+
+    indexed.loading = true
+
+    const value = getNotYetIndexed(key)
 
     index.set(key, {
       ...indexed,
-      subset,
+      value,
+      loading: false,
     })
+
+    return value
   }
-
-  return subset.get(...rest)
-}
-
-const createValueGetter = (
-  index,
-  getNotYetIndexed,
-  getSubset = createSubsetGetter(index, getNotYetIndexed),
-) => (key, ...rest) => {
-  const indexed = index.get(key) || {}
-
-  if (rest.length) {
-    return getSubset(indexed, key, rest)
-  }
-
-  if (indexed.loading) {
-    throw new Error('Cycle index referencing is not possible')
-  }
-
-  if ('value' in indexed) {
-    return indexed.value
-  }
-
-  indexed.loading = true
-
-  const value = getNotYetIndexed(key)
-
-  index.set(key, {
-    ...indexed,
-    value,
-    loading: false,
-  })
-
-  return value
-}
 
 export const createIndex = (getNotYetIndexed) => {
   let index = new Map()
