@@ -1,10 +1,15 @@
 import { hrToMs } from './hrToMs'
 
-const hasPerformance = typeof performance !== 'undefined'
+const hasPerformance =
+  typeof window !== 'undefined' && typeof performance !== 'undefined'
+
 const hasHr = typeof process !== 'undefined' && !!process.hrtime
 
 const dateRef =
-  (hasPerformance && performance.timing.navigationStart) || +new Date()
+  (hasPerformance &&
+    (performance.timeOrigin ||
+      (performance.timing && performance.timing.navigationStart))) ||
+  +new Date()
 
 const timeDeltaRef = hasHr && process.hrtime()
 
